Fix status bar and bottom sheet startup state
- Disable drag on bottom sheet
