+++ app/src/main/java/com/dcrandroid/fragments/SpendingPasswordFragment.kt
@@ -1,4 +1,10 @@
-package com.dcrandroid.ui.security
+/*
+ * Copyright (c) 2018-2019 The Decred developers
+ * Use of this source code is governed by an ISC
+ * license that can be found in the LICENSE file.
+ */
+
+package com.dcrandroid.fragments
 
 import android.os.Bundle
 import android.text.Editable
@@ -9,37 +15,36 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.dcrandroid.R
 import com.dcrandroid.util.Utils
-import kotlinx.android.synthetic.main.fragment_spending_pin.*
-import kotlinx.android.synthetic.main.passcode.pin_strength as pin_strength1
+import kotlinx.android.synthetic.main.fragment_spending_password.*
+import kotlinx.android.synthetic.main.password.password_strength
 
 
-class SpendingPinFragment(private var clickListener: DialogButtonListener) : Fragment() {
+class SpendingPasswordFragment(private var clickListener: DialogButtonListener) : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
                               savedInstanceState: Bundle?): View? {
-        return inflater.inflate(R.layout.fragment_spending_pin, container, false)
+        return inflater.inflate(R.layout.fragment_spending_password, container, false)
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        ed_pin.addTextChangedListener(pinWatcher)
-        ed_pin.addTextChangedListener(passwordStrengthWatcher)
-        ed_confirm_pin.addTextChangedListener(pinWatcher)
+        ed_password.addTextChangedListener(passwordWatcher)
+        ed_password.addTextChangedListener(passwordStrengthWatcher)
+        ed_confirm_password.addTextChangedListener(passwordWatcher)
 
         btn_cancel.setOnClickListener { clickListener.onClickCancel() }
         btn_create.setOnClickListener {
             it.visibility = View.GONE
-            ed_pin.isFocusable = false
-            ed_confirm_pin.isFocusable = false
-
+            ed_password.isFocusable = false
+            ed_confirm_password.isFocusable = false
             activity?.runOnUiThread {
                 btn_cancel.isEnabled = false
                 btn_cancel.setTextColor(resources.getColor(R.color.colorDisabled))
                 progress_bar.visibility = View.VISIBLE
             }
 
-            clickListener.onClickOk(ed_pin.text.toString())
+            clickListener.onClickOk(ed_password.text.toString())
         }
     }
 
@@ -51,37 +56,38 @@ class SpendingPinFragment(private var clickListener: DialogButtonListener) : Fra
         override fun afterTextChanged(s: Editable?) {
             val progress = (Utils.getShannonEntropy(s.toString()) / 4) * 100
             if (progress > 70) {
-                pin_strength.progressDrawable = resources.getDrawable(R.drawable.password_strength_bar_strong)
+                password_strength.progressDrawable = resources.getDrawable(R.drawable.password_strength_bar_strong)
             } else {
-                pin_strength.progressDrawable = resources.getDrawable(R.drawable.password_strength_bar_weak)
+                password_strength.progressDrawable = resources.getDrawable(R.drawable.password_strength_bar_weak)
             }
 
-            pin_strength.progress = progress.toInt()
+            password_strength.progress = progress.toInt()
         }
     }
 
-    private val pinWatcher = object : TextWatcher {
+    private val passwordWatcher = object : TextWatcher {
         override fun beforeTextChanged(s: CharSequence, start: Int, count: Int, after: Int) {}
 
         override fun onTextChanged(s: CharSequence, start: Int, before: Int, count: Int) {}
 
         override fun afterTextChanged(editable: Editable) {
-            val pinCount = ed_pin.text?.count()
-            val confirmPinCount = ed_confirm_pin.text?.count()
-            tv_confirm_pin_count.text = confirmPinCount.toString()
-            tv_pin_count.text = pinCount.toString()
+            val passwordCount = ed_password.text?.count()
+            val confirmPasswordCount = ed_confirm_password.text?.count()
+            tv_confirm_password_count.text = confirmPasswordCount.toString()
+            tv_password_count.text = passwordCount.toString()
 
-            if (ed_confirm_pin.text.toString() == "") {
-                til_confirm_pin.error = null
+            if (ed_confirm_password.text.toString() == "") {
+                til_confirm_password.error = null
             } else {
-                if (ed_pin.text.toString() != ed_confirm_pin.text.toString()) {
-                    til_confirm_pin.error = getString(R.string.mismatch_passcode)
+                if (ed_password.text.toString() != ed_confirm_password.text.toString()) {
+                    til_confirm_password.error = getString(R.string.mismatch_password)
                     btn_create.isEnabled = false
                 } else {
-                    til_confirm_pin.error = null
+                    til_confirm_password.error = null
                     btn_create.isEnabled = true
                 }
             }
         }
     }
+
 }
