--- app/src/main/java/com/dcrandroid/ui/security/DialogButtonListener.kt
@@ -1,12 +1,0 @@
-package com.dcrandroid.ui.security
-
-/**
- * Created by cheruiyot on 2019-07-29.
- */
-
-interface DialogButtonListener {
-
-    fun onClickOk(spendingKey: String)
-
-    fun onClickCancel()
-}
