+++ app/src/main/res/drawable/unselected_tab_bottom_indicator.xml
@@ -1,4 +1,5 @@
-<?xml version="1.0" encoding="utf-8"?><!--
+<?xml version="1.0" encoding="utf-8"?>
+<!--
     ~ Copyright (c) 2018-2019 The Decred developers
     ~ Use of this source code is governed by an ISC
     ~ license that can be found in the LICENSE file.
