+++ app/src/main/res/layout/fragment_password_pin_dialog.xml
@@ -1,4 +1,5 @@
-<?xml version="1.0" encoding="utf-8"?><!--
+<?xml version="1.0" encoding="utf-8"?>
+<!--
   ~ Copyright (c) 2018-2019 The Decred developers
   ~ Use of this source code is governed by an ISC
   ~ license that can be found in the LICENSE file.
@@ -9,6 +10,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:background="@color/white"
+    app:layout_behavior="android.support.design.widget.BottomSheetBehavior"
     android:orientation="vertical">
 
     <androidx.appcompat.widget.AppCompatTextView
