+++ app/src/main/res/layout/fragment_spending_pin.xml
@@ -1,4 +1,5 @@
-<?xml version="1.0" encoding="utf-8"?><!--
+<?xml version="1.0" encoding="utf-8"?>
+<!--
   ~ Copyright (c) 2018-2019 The Decred developers
   ~ Use of this source code is governed by an ISC
   ~ license that can be found in the LICENSE file.
@@ -26,6 +27,8 @@
         android:hint="@string/spending_pin"
         android:textColorHint="#3d5873"
         app:fontFamily="@font/source_sans_pro_regular_family"
+        app:boxStrokeColor="#2970ff"
+
         app:passwordToggleEnabled="true">
 
         <com.google.android.material.textfield.TextInputEditText
