+++ app/src/main/res/values/colors.xml
@@ -1,4 +1,5 @@
-<?xml version="1.0" encoding="utf-8"?><!--
+<?xml version="1.0" encoding="utf-8"?>
+<!--
   ~ Copyright (c) 2018-2019 The Decred developers
   ~ Use of this source code is governed by an ISC
   ~ license that can be found in the LICENSE file.
@@ -17,7 +18,7 @@
     <color name="colorAccent">#888a8b</color>
     <color name="lightGray">#C4CBD2</color>
     <color name="colorPrimary">#F3F5F6</color>
-    <color name="colorPrimaryDark">#e6eaed</color>
+    <color name="colorPrimaryDark">#838383</color>
     <color name="white">#ffffff</color>
     <color name="blue">#2970FF</color>
     <color name="green">#1dc54c</color>
